+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -6,8 +6,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handleradapter.HandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerMapping;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 import java.util.ArrayList;
